Combine the two Utilities classes
