+++ AAexam/app/src/main/java/sorting/LevelSort.java
@@ -23,7 +23,7 @@ public class LevelSort {
         }
         Comparable[] aux = new Comparable[a.length];
         int compares = sort(a, aux, c);
-        assert SortUtils.isSorted(a);
+        assert Util.isSorted(a);
         return compares;
     }
