+++ AAexam/app/src/test/java/sorting/BottomUpMergeSortCutoffTest.java
@@ -1,3 +1,5 @@
+package sorting;
+
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -6,9 +8,6 @@ import org.junit.jupiter.api.Test;
 
 import data.Handler;
 import data.TestData;
-import sorting.BottomUpMergeSort;
-import sorting.BottomUpMergeSortCutoff;
-import sorting.InsertionSort;
 
 public class BottomUpMergeSortCutoffTest {
