+++ AAexam/app/src/test/java/sorting/TopDownMergeSortTest.java
@@ -1,15 +1,34 @@
+package sorting;
+
+import static data.Handler.randomString;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static sorting.SortUtils.isSorted;
 
+import java.util.Arrays;
 
 import org.junit.jupiter.api.Test;
 
 import data.Handler;
 import data.TestData;
-import sorting.BottomUpMergeSort;
+import sorting.TopDownMergeSort;
 
-public class BottomUpMergeSortTest {
-    
+public class TopDownMergeSortTest {
+
+    // ====================================
+    // String tests
+    // ====================================
+
+    @Test void
+    givenRandomStrings_whenSort_thenSorted() {
+        int n = 100, m = 20;
+        String[] in = Handler.generate(n, i -> randomString(m));
+        assertFalse(isSorted(in));
+        TopDownMergeSort.sort(in);
+        assertTrue(isSorted(in));
+    }
 
     //-----------------------------------
     //Case: Empty array
@@ -17,16 +36,15 @@ public class BottomUpMergeSortTest {
     @Test
     void givenEmptyArray_whenMergeSort_thenEmptyArray() {
         Integer[] emptyArray = new Integer[0];
-        BottomUpMergeSort.sort(emptyArray);
+        TopDownMergeSort.sort(emptyArray);
         assertArrayEquals(new Integer[]{}, emptyArray);
     }
 
     @Test
     void givenEmptyArray_whenMergeSort_thenReturnZeroCompares() {
         Integer[] emptyArray = new Integer[0];
-        assertEquals(0, BottomUpMergeSort.sort(emptyArray));
+        assertEquals(0, TopDownMergeSort.sort(emptyArray));
     }
-
 
     //-----------------------------------
     //Case: Only one element
@@ -34,35 +52,31 @@ public class BottomUpMergeSortTest {
     @Test
     void givenSizeOneArray_whenMergeSort_thenIdenticalArray() {
         Integer[] sizeOneArray = new Integer[]{1};
-        BottomUpMergeSort.sort(sizeOneArray);
+        TopDownMergeSort.sort(sizeOneArray);
         assertArrayEquals(new Integer[]{1}, sizeOneArray);
     }
 
     @Test
     void givenSizeOneArray_whenMergeSort_thenReturnZeroCompares() {
         Integer[] sizeOneArray = new Integer[]{1};
-        assertEquals(0, BottomUpMergeSort.sort(sizeOneArray));
+        assertEquals(0, TopDownMergeSort.sort(sizeOneArray));
     }
 
-     //-----------------------------------
+    //-----------------------------------
     //Case: Descending order rather than ascending
-    //Also: just one run in stack after initial computation of runs
-    //------------------------------------
+    //-----------------------------------
     @Test
     void givenDescendingArray_whenMergeSort_thenAscendingArray() {
         Integer[] descendingArray = new Integer[]{2,1,0,-1};
-        BottomUpMergeSort.sort(descendingArray);
+        TopDownMergeSort.sort(descendingArray);
         assertArrayEquals(new Integer[]{-1,0,1,2}, descendingArray);
     }
 
     @Test
     void givenDescendingArray_whenMergeSort_thenReturnNumberOfCompares() {
         Integer[] descendingArray = new Integer[]{2,1,0,-1};
-        assertEquals(4, BottomUpMergeSort.sort(descendingArray));
+        assertEquals(4, TopDownMergeSort.sort(descendingArray));
     }
-
-    
-
 
     //-----------------------------------
     //Case: Odd amount of elements
@@ -70,32 +84,14 @@ public class BottomUpMergeSortTest {
     @Test
     void givenOddSizeArray_whenMergeSort_thenSortsArray() {
         Integer[] oddSizeArray = new Integer[]{9,1,2,0,7,4,3,8,5,6,10};
-        BottomUpMergeSort.sort(oddSizeArray);
+        TopDownMergeSort.sort(oddSizeArray);
         assertArrayEquals(new Integer[]{0,1,2,3,4,5,6,7,8,9,10}, oddSizeArray);
     }
 
     @Test
     void givenOddSizeArray_whenMergeSort_thenReturnNumberOfCompares() {
         Integer[] oddSizeArray = new Integer[]{9,1,2,0,7,4,3,8,5,6,10};
-        assertEquals(30, BottomUpMergeSort.sort(oddSizeArray));
-    }
-
-
-    //-----------------------------------
-    //Case: Seven elements (first three bits set in stack after initial merging of runs)
-    //-----------------------------------
-
-    @Test
-    void givenOddSize7Array_whenMergeSort_thenSortsArray() {
-        Integer[] oddSizeArray = new Integer[]{7,6,5,4,3,2,1};
-        BottomUpMergeSort.sort(oddSizeArray);
-        assertArrayEquals(new Integer[]{1,2,3,4,5,6,7}, oddSizeArray);
-    }
-
-    @Test
-    void givenOddSize7Array_whenMergeSort_thenReturnNumberOfCompares() {
-        Integer[] oddSizeArray = new Integer[]{7,6,5,4,3,2,1};
-        assertEquals(9, BottomUpMergeSort.sort(oddSizeArray));
+        assertEquals(29, TopDownMergeSort.sort(oddSizeArray));
     }
 
 
@@ -106,16 +102,15 @@ public class BottomUpMergeSortTest {
     @Test
     void givenIdenticalElementsArray_whenMergeSort_thenPreserveStability() {
         TestData[] identicalElementsArray = Handler.generate(10, i -> new TestData(i,0));
-        BottomUpMergeSort.sort(identicalElementsArray);
+        TopDownMergeSort.sort(identicalElementsArray);
         assertArrayEquals(Handler.generate(10, i -> new TestData(i,0)), identicalElementsArray);
     }
 
     @Test
     void givenIdenticalElementsArray_whenMergeSort_thenReturnNumberOfCompares() {
         TestData[] identicalElementsArray = Handler.generate(10, i -> new TestData(i,0));
-        assertEquals(21, BottomUpMergeSort.sort(identicalElementsArray));
+        assertEquals(19, TopDownMergeSort.sort(identicalElementsArray));
     }
-
 
     //-----------------------------------
     //Case: small array contains duplicates (for easier readability compared to medium array)
@@ -126,7 +121,7 @@ public class BottomUpMergeSortTest {
         TestData::from); 
         TestData[] expectedArray = Handler.readData(Handler.streamFile("unittest/duplicateElems.TestData.out"), 
         TestData::from); 
-        BottomUpMergeSort.sort(duplicateElementsArray);
+        TopDownMergeSort.sort(duplicateElementsArray);
         assertArrayEquals(expectedArray, duplicateElementsArray);
     }
 
@@ -134,9 +129,8 @@ public class BottomUpMergeSortTest {
     void givenDuplicateElementsArray_whenMergeSort_thenReturnNumberOfCompares() {
         TestData[] duplicateElementsArray = Handler.readData(Handler.streamFile("unittest/duplicateElems.TestData.in"), 
         TestData::from); 
-        assertEquals(9, BottomUpMergeSort.sort(duplicateElementsArray));
+        assertEquals(8, TopDownMergeSort.sort(duplicateElementsArray));
     }
-
 
     //-----------------------------------
     //Case: medium array contains duplicates
@@ -149,7 +143,16 @@ public class BottomUpMergeSortTest {
         TestData::from); 
         TestData[] expectedArray = Handler.readData(Handler.streamFile("unittest/duplicateElemsMedium.TestData.out"), 
         TestData::from); 
-        BottomUpMergeSort.sort(duplicateElementsArray);
+
+        // Debug: Log input and expected output
+        System.out.println("Input Array: " + Arrays.toString(duplicateElementsArray));
+        System.out.println("Expected Array: " + Arrays.toString(expectedArray));
+
+        TopDownMergeSort.sort(duplicateElementsArray);
+
+        // Debug: sorted input
+        System.out.println("Sorted Array: " + Arrays.toString(duplicateElementsArray));
+
         assertArrayEquals(expectedArray, duplicateElementsArray);
     }    
 
@@ -157,9 +160,8 @@ public class BottomUpMergeSortTest {
     void givenDuplicateElementsArrayMedium_whenMergeSort_thenReturnNumberOfCompares() {
         TestData[] duplicateElementsArray = Handler.readData(Handler.streamFile("unittest/duplicateElemsMedium.TestData.in"), 
         TestData::from); 
-        assertEquals(25, BottomUpMergeSort.sort(duplicateElementsArray));
+
+        assertEquals(23, TopDownMergeSort.sort(duplicateElementsArray));
     }    
-
-
 
 }
